Fix some deckbrew bugs.
First, it seems that we incorrectly make the URL with multiple '?'
signs. Secondly, sometimes the muid is 0, which causes deckbrew to
return all of its cards (I think), but then it's capped to some limit,
so we end up not getting all the cards we want, just those first in
alphabetical order.